#25 Rename components for consistent naming
