+++ frontend/src/pages/RegisterPage.tsx
@@ -1,4 +1,4 @@
-import { RegisterForm } from "@/components/register-form";
+import { RegisterForm } from "@/components/RegisterForm";
 
 export default function RegisterPage() {
     return (
